Consolidate ThemeWriter & ThemeGenerator & subclasses
