+++ hugo/src/main/scala/com/ossuminc/riddl/hugo/diagrams/mermaid/UseCaseDiagramSupport.scala
@@ -16,10 +16,10 @@ import scala.reflect.ClassTag
   */
 trait UseCaseDiagramSupport {
   def passesResult: PassesResult
+  def makeDocLink(definition: NamedValue): String
+
   def getDefinitionFor[T <: Definition: ClassTag](pathId: PathIdentifier, parent: Definition): Option[T] = {
     passesResult.refMap.definitionOf[T](pathId, parent)
   }
-  def makeDocLink(definition: NamedValue): String
+
 }
-
-
