+++ hugo/src/main/scala/com/ossuminc/riddl/hugo/writers/ContextWriter.scala
@@ -16,7 +16,7 @@ trait ContextWriter { this: MarkdownWriter =>
 
   def emitContext(context: Context, parents: Seq[Definition]): Unit = {
     containerHead(context, "Context")
-    val maybeDiagram = diagrams.contextDiagrams.get(context).map(data => mermaid.ContextMapDiagram(context, data))
+    val maybeDiagram = generator.diagrams.contextDiagrams.get(context).map(data => mermaid.ContextMapDiagram(context, data))
     emitVitalDefinitionDetails(context, parents)
     emitContextMap(context, maybeDiagram)
     emitOptions(context.options)
